authorization: removing a remnant from a merge conflict
These tests have been moved up a level, so this PR removes the `tests` folder
