+++ azurerm/internal/services/authorization/role_definition_resource_test.go
@@ -11,35 +11,12 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance/check"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/authorization/parse"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
 type RoleDefinitionResource struct{}
 
-func TestAccRoleDefinition(t *testing.T) {
-	acceptance.RunTestsInSequence(t, map[string]map[string]func(t *testing.T){
-		"basic": {
-			"basic":            testAccRoleDefinition_basic,
-			"management_group": testAccRoleDefinition_managementGroup,
-		},
-		"complete": {
-			"complete": testAccRoleDefinition_complete,
-		},
-		"empty": {
-			"empty_name":      testAccRoleDefinition_emptyName,
-			"update_empty_id": testAccRoleDefinition_updateEmptyId,
-		},
-		"import": {
-			"requiresImport": testAccRoleDefinition_requiresImport,
-		},
-		"update": {
-			"update": testAccRoleDefinition_update,
-		},
-	})
-}
-
-func testAccRoleDefinition_basic(t *testing.T) {
+func TestAccAzureRMRoleDefinition_basic(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 
@@ -54,10 +31,11 @@ func testAccRoleDefinition_basic(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_requiresImport(t *testing.T) {
+func TestAccAzureRMRoleDefinition_requiresImport(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
-	r := RoleDefinitionResource{}
 	id := uuid.New().String()
+
+	r := RoleDefinitionResource{}
 
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
@@ -72,7 +50,7 @@ func testAccRoleDefinition_requiresImport(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_complete(t *testing.T) {
+func TestAccAzureRMRoleDefinition_complete(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 
@@ -87,7 +65,7 @@ func testAccRoleDefinition_complete(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_update(t *testing.T) {
+func TestAccAzureRMRoleDefinition_update(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 	id := uuid.New().String()
@@ -101,7 +79,7 @@ func testAccRoleDefinition_update(t *testing.T) {
 		},
 		data.ImportStep(),
 		{
-			Config: r.basicUpdated(id, data),
+			Config: r.updated(id, data),
 			Check: resource.ComposeTestCheckFunc(
 				check.That(data.ResourceName).ExistsInAzure(r),
 			),
@@ -110,8 +88,9 @@ func testAccRoleDefinition_update(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_updateEmptyId(t *testing.T) {
+func TestAccAzureRMRoleDefinition_updateEmptyId(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
+
 	r := RoleDefinitionResource{}
 
 	data.ResourceTest(t, r, []resource.TestStep{
@@ -132,7 +111,7 @@ func testAccRoleDefinition_updateEmptyId(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_emptyName(t *testing.T) {
+func TestAccAzureRMRoleDefinition_emptyName(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 
@@ -147,13 +126,13 @@ func testAccRoleDefinition_emptyName(t *testing.T) {
 	})
 }
 
-func testAccRoleDefinition_managementGroup(t *testing.T) {
+func TestAccAzureRMRoleDefinition_managementGroup(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
-			Config: r.TestAccRoleDefinition_managementGroup(uuid.New().String(), data),
+			Config: r.managementGroup(uuid.New().String(), data),
 			Check: resource.ComposeTestCheckFunc(
 				check.That(data.ResourceName).ExistsInAzure(r),
 			),
@@ -162,7 +141,7 @@ func testAccRoleDefinition_managementGroup(t *testing.T) {
 	})
 }
 
-func TestAccRoleDefinition_assignToSmallerScope(t *testing.T) {
+func TestAccAzureRMRoleDefinition_assignToSmallerScope(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 	r := RoleDefinitionResource{}
 
@@ -177,7 +156,7 @@ func TestAccRoleDefinition_assignToSmallerScope(t *testing.T) {
 	})
 }
 
-func TestAccRoleDefinition_noAssignableScope(t *testing.T) {
+func TestAccAzureRMRoleDefinition_noAssignableScope(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_role_definition", "test")
 
 	r := RoleDefinitionResource{}
@@ -193,23 +172,22 @@ func TestAccRoleDefinition_noAssignableScope(t *testing.T) {
 	})
 }
 
-func (r RoleDefinitionResource) Exists(ctx context.Context, client *clients.Client, state *terraform.InstanceState) (*bool, error) {
-	id, err := parse.RoleDefinitionId(state.ID)
-	if err != nil {
-		return nil, err
-	}
+func (RoleDefinitionResource) Exists(ctx context.Context, client *clients.Client, state *terraform.InstanceState) (*bool, error) {
+	scope := state.Attributes["scope"]
+	roleDefinitionId := state.Attributes["role_definition_id"]
 
-	resp, err := client.Authorization.RoleDefinitionsClient.Get(ctx, id.Scope, id.RoleID)
+	resp, err := client.Authorization.RoleDefinitionsClient.Get(ctx, scope, roleDefinitionId)
 	if err != nil {
 		if utils.ResponseWasNotFound(resp.Response) {
 			return utils.Bool(false), nil
 		}
-		return nil, fmt.Errorf("retrieving Role Definition %q (Scope %q): %+v", id.RoleID, id.Scope, err)
+		return nil, fmt.Errorf("Bad: Get on roleDefinitionsClient: %+v", err)
 	}
+
 	return utils.Bool(true), nil
 }
 
-func (r RoleDefinitionResource) basic(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) basic(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -236,6 +214,7 @@ resource "azurerm_role_definition" "test" {
 }
 
 func (r RoleDefinitionResource) requiresImport(id string, data acceptance.TestData) string {
+	template := r.basic(id, data)
 	return fmt.Sprintf(`
 %s
 
@@ -253,10 +232,10 @@ resource "azurerm_role_definition" "import" {
     data.azurerm_subscription.primary.id,
   ]
 }
-`, r.basic(id, data))
+`, template)
 }
 
-func (r RoleDefinitionResource) complete(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) complete(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -285,7 +264,7 @@ resource "azurerm_role_definition" "test" {
 `, id, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) basicUpdated(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) updated(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -312,7 +291,7 @@ resource "azurerm_role_definition" "test" {
 `, id, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) emptyId(data acceptance.TestData) string {
+func (RoleDefinitionResource) emptyId(data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -337,7 +316,7 @@ resource "azurerm_role_definition" "test" {
 `, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) updateEmptyId(data acceptance.TestData) string {
+func (RoleDefinitionResource) updateEmptyId(data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -362,7 +341,7 @@ resource "azurerm_role_definition" "test" {
 `, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) TestAccRoleDefinition_managementGroup(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) managementGroup(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -392,7 +371,7 @@ resource "azurerm_role_definition" "test" {
 `, id, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) assignToSmallerScope(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) assignToSmallerScope(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
@@ -423,7 +402,7 @@ resource "azurerm_role_definition" "test" {
 `, data.RandomInteger, data.Locations.Primary, id, data.RandomInteger)
 }
 
-func (r RoleDefinitionResource) noAssignableScope(id string, data acceptance.TestData) string {
+func (RoleDefinitionResource) noAssignableScope(id string, data acceptance.TestData) string {
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
